requirements-minimal.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
Pillow==10.0.1
requests==2.31.0
cloudinary==1.36.0
python-dotenv==1.0.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
cmake
dlib-bin==19.24.2
face-recognition==1.3.0